Added A Flickering Light Script and Started Object pickup in the the Third Person Blue print
Flickering Light script is called BlinkingLight and is point light that
delays visibility. More delays can be added very easily and more realism
is added with every delay. I scripted a large chunk of the Object pickup
mechanic in the our character controller. Feel free to add to it or fix
the wrong behavior that it has now.