Added events up to the tv. can't turn it off for some reason.
still working on it. had to commit so LA could work. Added a couple
essential props such as pill bottles for the bathroom to simulate being
messed up and the vase we were gonna have dead flowers in. there's also
a few more generic noises going on. sorry about the tv static